+++ src/functions/beautifyPerms.js
@@ -4,13 +4,13 @@
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
 
-module.exports = function (perms, allPerms) {
+module.exports = function (perms, allPerms, t) {
     let index;
     let newPerms = [];
     for (var i = 0; i < allPerms.length; i++) {
         index = perms.indexOf(allPerms[i].perm);
         if (index !== -1) {
-            newPerms[index] = allPerms[i].val;
+            newPerms[index] = t(`permissions:${allPerms[i].val}`);
         }
     }
     return newPerms;